sample-app/requirements.txt
azure-storage-blob==12.18.3
azure-identity==1.15.0
kubernetes==30.1.0